requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
whatsapp-api-client-python==0.0.53
httpx==0.27.2
python-dotenv==1.0.1
tenacity==9.0.0
anthropic>=0.86.0
websockets>=12.0
certifi
